+++ mod.cpp
@@ -1,6 +1,5 @@
-name = "Weapons balance - Project BLUFOR";
-description = "A collection of weight balancing mods made by johnb43";
-tooltip = "Weight balance";
-tooltipOwned = "Weight balance";
-overview = "";
+name = "Weapons Balance - Project Blufor";
+description = "A collection of balancing mods made by johnb43";
+tooltip = "Weapons Balance - Project Blufor";
+tooltipOwned = "Weapons Balance - Project Blufor";
 author = "johnb43";